esm: move package config helpers
